requirements.txt
Flask==3.1.1
pdfplumber==0.11.7
python-docx==1.2.0
google-generativeai==0.8.5
docxcompose==1.4.0
